:truck: move theme-provider to layouts dir
seems a fairly better place for it
